Trabalhando no projeto

+++ src/main/java/com/livro/api/livroapi/service/AuthorService.java
@@ -1,21 +1,15 @@
 package com.livro.api.livroapi.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.livro.api.livroapi.model.Author;
 import com.livro.api.livroapi.dto.AuthorDTO;
 import com.livro.api.livroapi.repository.AuthorRepository;
-import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import org.springframework.beans.BeanUtils;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -41,15 +35,20 @@ public class AuthorService {
     }
 	
 	@Transactional
-	public void update(String uuid, Author author) {
+	public void update(String uuid, AuthorDTO authorDTO) {
 		Author authorAtual = authorRepository.findById(UUID.fromString(uuid)).get();
-		BeanUtils.copyProperties(author, authorAtual, "id");
+		BeanUtils.copyProperties(authorDTO, authorAtual, "id");
 		authorRepository.save(authorAtual);
 	}
 	
 	@Transactional
-	public void delete(String uuid) {
-		authorRepository.deleteById(UUID.fromString(uuid));
+	public boolean delete(String uuid) {
+		Optional<Author> author = authorRepository.findById(UUID.fromString(uuid));
+		if(author.isPresent()) {
+			authorRepository.deleteById(UUID.fromString(uuid));
+			return true;
+		}
+		return false;
 	}
 	
 	public List<Author> getAll(String orderBy) {
